Update infinite-scrolling.mdx (#117)
I was a little confused by the wording and meaning of the first sentence, so I moved the word "that" so that it would make more sense to me. But please make sure that it's correct. I figure there was a typo and wasn't sure what you had wanted it to say, so I just took my best guess.

Also, I noticed that in other places, you referred to the virtualization feature as a singular feature, so I changed it to singular here, but feel free to change it back to what you had before.