Update column-actions.mdx
One change I might wan to draw your attention is on line 51 of the code, where I added the word "for" after the word "property." I might have misread this sentence, but when I read it, I thought it needed another word. But, like I said, I might have misunderstood the meaning of the sentence.