Range slider, array of arrays from getFacetedMinMaxValues() (#570)
This commit implements a simple array check for the values returned by getFacetedMinMaxValues(). The bug was encountered while using range-slider for column that had two rows, and the first row had a smaller value in the column which was controlled by the range slider. In this configuration, the min value of the result from getFacetedMinMaxValues was an array, when a number was expected. Pretty much any other configuration seemed to work, but since this pretty hard to know for sure, the same check is implemented for the max value too.